# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: MALIGAIN API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /wallet/transactions:
    get:
      operationId: getTransactions
      tags: [wallet]
      summary: Get transaction history
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [withdrawal, recharge, all]
            default: all
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Transaction list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionList"
...
  /admin/auth/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminAuthResponse"
        "401":
          description: Invalid credentials
...
  /admin/auth/me:
    get:
      operationId: adminGetMe
      tags: [admin]
      summary: Get current admin
      responses:
        "200":
          description: Current admin user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
        "401":
          description: Not authenticated
...
  /admin/transactions:
    get:
      operationId: adminListTransactions
      tags: [admin]
      summary: List all transactions platform-wide
      parameters:
        - name: userId
          in: query
          schema:
            type: integer
        - name: type
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
      responses:
        "200":
          description: Transaction list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminTransactionList"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
...
    LoginRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
        rememberMe:
          type: boolean
          default: false
      required:
        - email
        - password
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        message:
          type: string
      required:
        - user
        - message
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        phone:
          type: string
          nullable: true
        country:
          type: string
          nullable: true
        referralCode:
          type: string
        status:
          type: string
          enum: [active, inactive, suspended]
        avatarInitials:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - username
        - email
        - referralCode
        - status
        - avatarInitials
        - createdAt
...
    Transaction:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [withdrawal, recharge, bonus, commission, referral]
        amount:
          type: number
        status:
          type: string
          enum: [pending, completed, failed]
        description:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - type
        - amount
        - status
        - description
        - createdAt
    TransactionList:
      type: object
      properties:
        transactions:
          type: array
          items:
            $ref: "#/components/schemas/Transaction"
        total:
          type: integer
        totalPaidOut:
          type: number
        totalPending:
          type: integer
      required:
        - transactions
        - total
        - totalPaidOut
        - totalPending
...
    AdminLoginRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password
    AdminUser:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
      required:
        - id
        - username
    AdminAuthResponse:
      type: object
      properties:
        admin:
          $ref: "#/components/schemas/AdminUser"
        message:
          type: string
      required:
        - admin
        - message
...
    AdminTransactionItem:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        username:
          type: string
        type:
          type: string
        amount:
          type: number
        status:
          type: string
        description:
          type: string
          nullable: true
        phoneNumber:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - username
        - type
        - amount
        - status
        - createdAt
    AdminTransactionList:
      type: object
      properties:
        transactions:
          type: array
          items:
            $ref: "#/components/schemas/AdminTransactionItem"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required:
        - transactions
        - total
        - page
        - totalPages
...